fix: offload a lot from index to helpers change eodashboard -> eoxhub test catalog aligned with our actual live vitepress apps add storytelling dep
